refactor: organize fonts
